requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.3.4
openai==1.12.0
pdfplumber==0.10.3
python-docx==1.1.0
websockets==12.0